% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{make_phylo}
\alias{make_phylo}
\alias{read_newick_phylo}
\alias{read_nexus_phylo}
\title{Read phylogenetic trees from files}
\usage{
make_phylo(file, simplify = FALSE, ...)

read_nexus_phylo(file, simplify = FALSE, ...)

read_newick_phylo(file, simplify = FALSE, ...)
}
\arguments{
\item{file}{Path of NEXUS or Newick file}

\item{simplify}{If the file includes more than one tree, returns
only the first tree; otherwise, returns a multiPhylo object}

\item{...}{additional parameters to be passed to the rncl function}
}
\value{
A phylo or a multiPhyloo object
}
\description{
Create phylo objects from NEXUS or Newick files
}
\details{
These functions read NEXUS or Newick files and return an object of
class phylo/multiPhylo.
}
\note{
\code{make_phylo} may become deprecated in the future, use
\code{read_nexus_phylo} or \code{read_newick_phylo} instead.
}
\author{
Francois Michonneau
}
\seealso{
rncl, and the
}

