% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{readNCL}
\alias{readNCL}
\title{readNCL something}
\usage{
readNCL(file, simplify = FALSE, type = c("all", "tree", "data"),
  char.all = FALSE, polymorphic.convert = TRUE, levels.uniform = FALSE,
  quiet = TRUE, check.node.labels = c("keep", "drop", "asdata"),
  return.labels = TRUE, file.format = c("nexus", "newick"),
  check.names = TRUE, convert.edge.length = FALSE, ...)
}
\arguments{
\item{file}{something}

\item{simplify}{something}

\item{type}{something}

\item{char.all}{something}

\item{polymorphic.convert}{something}

\item{levels.uniform}{something}

\item{quiet}{something}

\item{check.node.labels}{something}

\item{return.labels}{something}

\item{file.format}{something}

\item{check.names}{something}

\item{convert.edge.length}{something}

\item{...}{something else}
}
\value{
something cool
}
\description{
Returns a list of the elements contained in a Nexus file used to build phylogenetic objects in R
}
\details{
The NEXUS class is a common file format
}

