% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{readNCL}
\alias{readNCL}
\title{readNCL something}
\usage{
readNCL(file, char.all = FALSE, polymorphic.convert = TRUE,
  levels.uniform = FALSE, quiet = TRUE, file.format = c("nexus",
  "newick"), check.names = TRUE, ...)
}
\arguments{
\item{file}{something}

\item{char.all}{something}

\item{polymorphic.convert}{something}

\item{levels.uniform}{something}

\item{quiet}{something}

\item{file.format}{something}

\item{check.names}{something}

\item{...}{something else}

\item{simplify}{something}

\item{type}{something}

\item{check.node.labels}{something}

\item{return.labels}{something}
}
\value{
something cool
}
\description{
Returns a list of the elements contained in a Nexus file used to
build phylogenetic objects in R
}
\details{
The NEXUS class is a common file format used in
phylogenetics. This function calls NCL to parse NEXUS, newick or
other common phylogenetic file formats.
}
\author{
Francois Michonneau
}

